Adjusting ignores for git
Reflect to fossil where feasible. In this case, tags/TAGS/etc might
need to be versioned, but aren't at the moment and are manually managed
until I add invocations into my Makefile that create them. We test for both
of cscope and exuberant ctags (not emacs ctags). This will need to be
propagated to both this branch and to master for both fossil and git.